数据库管理

+++ cloud/src/database/meta.rs
@@ -44,32 +44,32 @@ pub trait CloudMetaManager {
 pub trait FileManager {
     async fn list_deleted_file(&self, update_time: i64) -> Vec<FileMeta>;
 
-    async fn list_by_parent(&self, parent_id: i64) -> ResponseResult<Vec<FileMeta>>;
+    async fn list_by_parent(&self, parent_id: i32) -> ResponseResult<Vec<FileMeta>>;
 
-    async fn info_by_parent_and_name(&self, parent_id: i64, name: &str) -> Option<FileMeta>;
+    async fn info_by_parent_and_name(&self, parent_id: i32, name: &str) -> Option<FileMeta>;
 
     async fn new_file(
         &self,
-        parent_id: i64,
+        parent_id: i32,
         name: &str,
         file_type: FileMetaType,
     ) -> ResponseResult<FileMeta>;
 
     async fn update_meta(&self, meta: FileMeta) -> Option<FileMeta>;
     async fn update_file_content(&self, meta: FileMeta, block_index: usize) -> Option<FileMeta>;
-    async fn delete_file_blocks(&self, id: i64, block_index: i64);
+    async fn delete_file_blocks(&self, id: i32, block_index: i64);
 
-    async fn info_by_id(&self, id: i64) -> ResponseResult<Option<FileMeta>>;
+    async fn info_by_id(&self, id: i32) -> ResponseResult<Option<FileMeta>>;
 
-    async fn delete_file_meta(&self, id: i64) -> Option<FileMeta>;
+    async fn delete_file_meta(&self, id: i32) -> Option<FileMeta>;
 
-    async fn clean_file_meta(&self, id: i64) -> ResponseResult<Option<FileMeta>>;
+    async fn clean_file_meta(&self, id: i32) -> ResponseResult<Option<FileMeta>>;
 
-    async fn file_block_meta(&self, file_meta_id: i64) -> Vec<FileBlockMeta>;
+    async fn file_block_meta(&self, file_meta_id: i32) -> Vec<FileBlockMeta>;
 
-    async fn file_block_meta_index(&self, file_meta_id: i64, start: i64) -> Option<FileBlockMeta>;
+    async fn file_block_meta_index(&self, file_meta_id: i32, start: i64) -> Option<FileBlockMeta>;
 
-    async fn file_block_meta_info_by_id(&self, id: i64) -> Option<FileBlockMeta>;
+    async fn file_block_meta_info_by_id(&self, id: i32) -> Option<FileBlockMeta>;
 
     async fn update_file_block_meta(
         &self,
@@ -80,7 +80,7 @@ pub trait FileManager {
 
     async fn new_file_block_meta(
         &self,
-        file_meta_id: i64,
+        file_meta_id: i32,
         block_index: i64,
     ) -> Option<FileBlockMeta>;
